backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
playwright>=1.44.0
pdfminer.six>=20221105
requests>=2.31.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0